HR-app/backend/requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
pymongo==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing-extensions
tzdata==2024.1

# Added for AI Resume Screening
google-generativeai==0.7.2
pypdf2==3.0.1
python-docx==1.1.2
celery==5.4.0
redis==5.0.7 